노트 원 기울기 수정 템포가 마디단위로 설정이 아닌, Tick 단위로 설정되도록 변경 현재 Tick 계산을 plus tick을 더하는 방식에서 현재 시간을 Tick으로 변환시켜 대입하는 방식으로 변경

+++ KaraokeViewer/app/src/main/java/com/karaokepang/Activity/MusicListener.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 
 public interface MusicListener {
     public void notifyMeasureChanged(ArrayList<String> lyrics, long tick);
-    public void notifyCurrentTick(long tick, int term, int measureLength);
+    public void notifyCurrentTick(float tick, int term, int measureLength);
 }
